Move useful functions from desktopMain to skikoMain
