+++ compose/foundation/foundation/src/desktopMain/kotlin/androidx/compose/foundation/text/ContextMenu.desktop.kt
@@ -409,7 +409,6 @@ fun TextContextMenuArea(
     )
 }
 
-
 /**
  * The default text context menu items.
  *
